move integration internals into src/internal/
- src/internal/ holds the orchestration glue + vendored bits the
  user doesn't construct directly: turnkey{Claim,Transfer,Swap,
  Withdraw,Lightning,Internal}.ts and htlc-transactions.ts
- src/ now contains only entry-point scripts and the public
  TurnkeySparkSigner facade
- imports updated; typecheck clean
