+++ examples/with-spark/src/internal/htlc-transactions.ts
