+++ examples/with-spark/src/internal/turnkeyClaim.ts
@@ -22,7 +22,7 @@ import type {
   ClaimLeafInput,
   OperatorRecipientInput,
   TurnkeySparkSigner,
-} from "./turnkeySigner";
+} from "../turnkeySigner";
 import {
   type ClaimPackage,
   createSparkClient,
